requirements.txt
clodius>=0.10.3
higlass-python>=0.1.1
docker
ipywidgets
slugid==2.0.0
